backend/requirements.txt
fastapi==0.109.0
pillow==9.0.1
pydantic==1.10.11 
langchain==0.2.16
langchain-groq==0.1.10
langchain-core==0.2.43
groq ==0.18.0
beautifulsoup4==4.13.3
uvicorn==0.34.1
streamlit==1.30.0
requests==2.31.0
requests-toolbelt==1.0.0
python-dotenv==1.0.1
httpx==0.28.1
